@@ -4,14 +4,13 @@ version: 1.0.3
 homepage: https://github.com/FooStudio/tinycolor
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.0 <3.0.0"
 
 dependencies:
-  pigment: ^1.0.3
-  meta: ^1.2.2
   flutter:
     sdk: flutter
+  pigment: ^1.0.4
 
 dev_dependencies:
   flutter_test:
