@@ -2,13 +2,11 @@ import 'dart:math' as Math;
 import 'dart:ui';
 
 import 'package:flutter/painting.dart';
-import 'package:meta/meta.dart';
 
 import 'hsl_color.dart';
 import 'util.dart';
 
-HslColor rgbToHsl(
-    {@required double r, @required double g, @required double b}) {
+HslColor rgbToHsl({required double r, required double g, required double b}) {
   r = bound01(r, 255.0);
   g = bound01(g, 255.0);
   b = bound01(b, 255.0);
@@ -67,7 +65,7 @@ HSVColor colorToHsv(Color color) {
 }
 
 HSVColor rgbToHsv(
-    {@required int r, @required int g, @required int b, @required int a}) {
+    {required int r, required int g, required int b, required int a}) {
   return colorToHsv(Color.fromARGB(a, r, g, b));
 }
 
