@@ -2,7 +2,6 @@ import 'dart:math' as Math;
 import 'dart:ui';
 
 import 'package:flutter/painting.dart';
-import 'package:meta/meta.dart';
 import 'package:pigment/pigment.dart';
 
 import 'conversion.dart';
@@ -13,18 +12,15 @@ export 'hsl_color.dart';
 export 'color_extension.dart';
 
 class TinyColor {
-  Color originalColor;
+  final Color originalColor;
   Color _color;
 
-  TinyColor(Color color) {
-    this.originalColor =
-        Color.fromARGB(color.alpha, color.red, color.green, color.blue);
-    this._color =
-        Color.fromARGB(color.alpha, color.red, color.green, color.blue);
-  }
+  TinyColor(Color color)
+      : this.originalColor = color,
+        _color = color.copy();
 
   factory TinyColor.fromRGB(
-      {@required int r, @required int g, @required int b, int a = 100}) {
+      {required int r, required int g, required int b, int a = 100}) {
     return TinyColor(Color.fromARGB(a, r, g, b));
   }
 
@@ -142,7 +138,7 @@ class TinyColor {
     return TinyColor.fromHSL(hsl);
   }
 
-  TinyColor mix({@required Color input, int amount = 50}) {
+  TinyColor mix({required Color input, int amount = 50}) {
     final int p = (amount / 100).round();
     final color = Color.fromARGB(
         (input.alpha - _color.alpha) * p + _color.alpha,
@@ -160,5 +156,11 @@ class TinyColor {
 
   Color get color {
     return _color;
+  }
+}
+
+extension on Color {
+  Color copy() {
+    return Color.fromARGB(alpha, red, green, blue);
   }
 }
