@@ -4,7 +4,7 @@ class HslColor {
   double l;
   double a;
 
-  HslColor({this.h, this.s, this.l, this.a = 0.0});
+  HslColor({required this.h, required this.s, required this.l, this.a = 0.0});
 
   String toString() {
     return "HSL(h: $h, s: $s, l: $l, a: $a)";
